Implemented gas usage calculation in events

+++ tsa-core/src/main/kotlin/org/usvm/machine/state/TvmEventLogEntry.kt
@@ -2,7 +2,13 @@ package org.usvm.machine.state
 
 import org.usvm.machine.state.messages.ReceivedMessage
 
-typealias EventId = Long
+/**
+ * EventId is equal to the number of already executed instruction in state.
+ * It is unlikely that this number will exceed the boundaries of the Int type
+ * (at least for the reason we are storing all the gas usages in a collection
+ * that must have a size within Int boundaries).
+ */
+typealias EventId = Int
 
 /**
  * When smart contracts on TON network are executed, they follow the actor model ---
@@ -19,10 +25,15 @@ sealed interface TvmEventLogEntry {
     val id: EventId
 }
 
+/**
+ * This class is only reasonable within the context of the state where the entry occurred.
+ * For example, the gas usage history would be evaluated as
+ * `state.gasUsageHistory(executionBegin, executionEnd)`
+ */
 data class TvmMessageDrivenContractExecutionEntry(
     override val id: EventId,
-    val executionBegin: Long,
-    val executionEnd: Long,
+    val executionBegin: Int,
+    val executionEnd: Int,
     val contractId: ContractId,
     val incomingMessage: ReceivedMessage,
     val methodResult: TvmMethodResult,
